Fix getWeek for DST changes
A DST changes forwards produces a decimal value for the "days since
start of year" part of the calculation, which depending on the value of
"first day of week" might throw the calculation off. This fix rounds off
the decimal. Using round because rounding to the closes full day should
fix it in all cases as all DST changes are up to only 1 hour.